requirements.txt
# Core Dependencies (compatible versions)
langchain
langchain-core
langchain-google-genai
langchain-community
langchain-text-splitters

# Vector Database
chromadb

# LangGraph
langgraph

# FastAPI and Server
fastapi
uvicorn[standard]
python-multipart

# Utilities
pydantic>=2.0
pydantic-settings
python-dotenv

# Google AI
google-generativeai

# Additional utilities
tiktoken
requests

# Testing
pytest
pytest-asyncio
httpx
